Arrange constants directory
