--- src/constants/MarkdownEditor/editor_style.ts
@@ -1,13 +1,0 @@
-export const STYLE_MAP = {
-  BOLD: {
-    fontWeight: 'bold',
-  },
-
-  ITALIC: {
-    fontStyle: 'italic',
-  },
-
-  UNDERLINE: {
-    textDecoration: 'underline',
-  },
-}
